add OpCopy: 27.32s

+++ optimizer/sum_incr_decr.go
@@ -10,14 +10,14 @@ func SumIncrDecr(src []instr.Instruction) []instr.Instruction {
 			j := SearchCont(src, instr.OpIncr, i)
 			retval = append(retval, instr.Instruction{
 				Op:   instr.OpIncr,
-				Data: uint8(j - i),
+				Data: int16(j - i),
 			})
 			i = j - 1
 		} else if src[i].Op == instr.OpDecr {
 			j := SearchCont(src, instr.OpDecr, i)
 			retval = append(retval, instr.Instruction{
 				Op:   instr.OpDecr,
-				Data: uint8(j - i),
+				Data: int16(j - i),
 			})
 			i = j - 1
 		} else {
